workload支持浮点型

+++ src/global.cc
@@ -3,7 +3,7 @@
 int BLOCKS = 12;
 int THREADS = 256;
 
-int WORKLOAD_RATIO = 1; // 默认GPU/CPU任务比例1:1，即各自负责一半的串
+double WORKLOAD_RATIO = 1; // 默认GPU/CPU任务比例1:1，即各自负责一半的串
 
 int MODE = CPU_GPU;     // 默认同时使用GPU和CPU
